feat(suparun): v0.2.0
Admin tool (web):
- Config CRUD with Tabler UI, inline edit, auto-save, rewards modal
- Table read-only view with filter, pagination, stats, Chart.js histogram
- Cross-table search (user_id INTERSECT)
- Admin user management (DB-based roles, auto-registration)
- Audit log (all changes tracked with before/after snapshots)
- OAuth login support, XSS protection, import transaction safety

Features:
- SeasonPass (battle pass: free/premium tracks, XP levels)
- Config/Table group attribute: [Config("gacha")] sidebar grouping
- FK dropdown: foreignKey fields auto-render as <select>
- Ctrl+Z undo with toast feedback

Security & Performance:
- Reflection removed: typed delegate dictionaries for all CRUD
- XSS: escHtml() with quote escaping, audit JSON via array index
- Import: DeleteAll + SaveAll wrapped in transaction
- Field whitelist + parameter binding for Table queries

+++ com.tjdtjq5.suparun/Features~/Currency/CurrencyLog.cs
@@ -7,8 +7,8 @@ public class CurrencyLog
     [PrimaryKey] public string id;
     [Index] public string playerId;
     public string currencyId;
-    public int change;
-    public int balanceAfter;
+    public long change;
+    public long balanceAfter;
     [NotNull] public string reason;
     [CreatedAt] public long createdAt;
 }
